feat: change validade patient ownership as shared module
