+++ src/app/modules/Patient/module/index.ts
@@ -1,11 +1,11 @@
 import prisma from '../../../shared/database/prisma';
+import ValidatePatientOwnershipService from '../../../shared/services/ValidatePatientOwnership';
 import PatientController, { optionsController } from '../controller';
 import PatientRepositories from '../repositories/patient/PatientRepositories';
 import CreatePatientService from '../services/Create';
 import DeleteService from '../services/Delete';
 import FindAllPatientService from '../services/FindAll';
 import FindByPatientId from '../services/FindByUserId';
-import ValidatePatientOwnershipService from '../services/ValidatePatientOwnership';
 
 const patientRepositoriesInstance = new PatientRepositories(prisma);
