+++ src/app/modules/Patient/services/FindByUserId/FindByPatientId.ts
@@ -1,5 +1,5 @@
+import ValidatePatientOwnershipService from '../../../../shared/services/ValidatePatientOwnership';
 import PatientRepositories from '../../repositories/patient/PatientRepositories';
-import ValidatePatientOwnershipService from '../ValidatePatientOwnership';
 
 export class FindByPatientId {
   constructor(
