--- src/app/modules/Patient/services/ValidatePatientOwnership/index.ts
@@ -1,3 +1,0 @@
-import { ValidatePatientOwnershipService } from './ValidatePatientOwnershipService';
-
-export default ValidatePatientOwnershipService;
