+++ src/app/shared/services/ValidatePatientOwnership/index.ts
@@ -1,0 +1,3 @@
+import { ValidatePatientOwnershipService } from './ValidatePatientOwnershipService';
+
+export default ValidatePatientOwnershipService;
